mooclet_engine/requirements_base.txt
Django==1.11
requests
numpy>=1.10
djangorestframework
django-cors-middleware
django-filter
rest-pandas
sqlalchemy
coreapi
pygments
markdown
pycurl==7.43.0 --global-option="--with-nss"
django-celery-results
django-celery-beat